Blason doesn't overlap form when it's hidden
Before this commit the form was considered below the blason even when the blason
is transparent, the form was hence unusable.
With this commit the blason is hidden when it's not displayed.
Note that we need to add a sleep otherwise the smooth 2s transition isn't respected
(as if the browser finishes executing the script before updating the display, so it
 ends up applying both the change of opacity and the hidden status at once)